fix(RQF-21): Validação de Cod. Propriedade na tela de UP
Está ocorrendo erro quando o campo "Já possui código de propriedade?" está marcado com sim. Está retornando a validação "Código da propriedade superior ao máximo permitido.".

Estamos fazendo testes e revendo a lógica da validação - que parece estar errada, usando o valor mínimo 3250 como valor máximo na validação, sendo que ele é o mínimo (conforme entendimentos do João e Mateus) - e efetuando a correção do tipo de Int32 para Int64.

Além disso, corrigimos o campo Cod. Propriedade no Credenciado, de 4 para 11 dígitos.

+++ Código Fonte/VERSAO-ATUAL/src/Credenciado/Credenciado.Model.Extensoes/Caracterizacoes/ModuloUnidadeProducao/Business/UnidadeProducaoValidar.cs
@@ -65,7 +65,7 @@ namespace Tecnomapas.EtramiteX.Credenciado.Model.Extensoes.Caracterizacoes.Modul
 				}
 				else
 				{
-					if (caracterizacao.CodigoPropriedade > Convert.ToInt32(_configSys.Obter<String>(ConfiguracaoSistema.KeyUnidadeProducaoMaxCodigoPropriedade)))
+					if (caracterizacao.CodigoPropriedade > Convert.ToInt64(_configSys.Obter<String>(ConfiguracaoSistema.KeyUnidadeProducaoMaxCodigoPropriedade)))
 					{
 						Validacao.Add(Mensagem.UnidadeProducao.CodigoPropriedadeSuperiorMaximo);
 					}
